Fixed precision error in live value
Live numerical values were getting converted to floats which it some
cases led to precision loss. Changed to Double to resolve this issue.

+++ eWonConnector-gateway/src/main/java/org/imdc/ewon/SyncManager.java
@@ -291,7 +291,7 @@ public class SyncManager {
                         }
                         // Value is a number
                         else {
-                            value = Float.parseFloat(valueString);
+                            value = Double.parseDouble(valueString);
                         }
                     } catch (NullPointerException e) {
                         logger.error("Tag: " + tag + " does not exist on Ewon: " + eWonName, e);
